+++ src/App.js
@@ -3,8 +3,9 @@ import {Route, Switch} from 'react-router-dom';
 
 import Nav from './components/Nav';
 import Login from './components/Login/Login';
-import ClientPage from './components/Client/ClientForm';
+import ClientPage from './components/Client/ClientPage';
 import ClassForm from './components/Client/ClassForm';
+import InstructorPage from './components/Instructor/InstructorPage';
 
 function App() {
   return (
@@ -12,6 +13,7 @@ function App() {
       <Nav />
       Anytime. Anywhere. Fitness.
 
+      <Route path="/instructor" component={InstructorPage}/>
       <Route path="/classes" component={ClassForm}/>
       <Route path="/client">
         <ClientPage />
